+++ tests/_param_data/test_param_data.py
@@ -95,10 +95,10 @@ def test_child_does_not_change(param_data: ParamData[Any]) -> None:
 
 def test_to_and_from_json(param_data: ParamData[Any]) -> None:
     """Parameter data can be converted to and from JSON data."""
-    timestamp = param_data.last_updated.timestamp()
     json_data = param_data.to_json()
+    timestamp = param_data.last_updated.timestamp()
     with capture_start_end_times():
-        param_data_from_json = param_data.from_json(timestamp, json_data)
+        param_data_from_json = param_data.from_json(json_data, timestamp)
     assert param_data_from_json == param_data
     assert param_data_from_json.last_updated == param_data.last_updated